Initial Inventory Project

+++ src/main/java/com/usanase/AirtelInventorySystem/repository/DeviceRepository.java
@@ -3,5 +3,5 @@ package com.usanase.AirtelInventorySystem.repository;
 import com.usanase.AirtelInventorySystem.model.Device;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeviceRepository extends JpaRepository<Device,Integer> {
+public interface DeviceRepository extends JpaRepository<Device, Long> {
 }
